+++ truffle/src/com.oracle.truffle.api.dsl.test/src/com/oracle/truffle/api/dsl/test/CachedContextTest.java
@@ -465,6 +465,22 @@ public class CachedContextTest extends AbstractPolyglotTest {
         }
     }
 
+    abstract static class CachedLanguageError7Node extends Node {
+
+        abstract Object execute(Object argument);
+
+        /*
+         * Warning expected here as cachedEnv does not bind any dynamic parameter.
+         */
+        @ExpectError("The limit expression has no effect. %")
+        @Specialization(guards = "cachedEnv != null", limit = "3")
+        static String s0(Object value,
+                        @CachedContext(CachedContextTestLanguage.class) Env env,
+                        @Cached("env") Env cachedEnv) {
+            throw new AssertionError();
+        }
+    }
+
     @GenerateLibrary
     public abstract static class CachedContextTestLibrary extends Library {
 
@@ -483,7 +499,7 @@ public class CachedContextTest extends AbstractPolyglotTest {
         }
 
         @ExportMessage
-        final Object m1(@CachedContext(CachedContextTestLanguage.class) ContextReference<Env> env) {
+        final Object m1(@CachedContext(CachedContextTestLanguage.class) Env env) {
             return "m1";
         }
     }
@@ -501,6 +517,23 @@ public class CachedContextTest extends AbstractPolyglotTest {
         @Specialization
         Object doSomething(@CachedContext(CachedContextTestLanguage.class) Env ctx) {
             return null;
+        }
+    }
+
+    public abstract static class CacheCachedContextTest extends Node {
+
+        protected abstract boolean executeMatch(Object arg0);
+
+        @ExpectError("The limit expression has no effect.%")
+        @Specialization(guards = {"cachedGuard"}, limit = "1")
+        boolean s0(Object arg0,
+                        @CachedContext(CachedContextTestLanguage.class) Env env,
+                        @Cached("getBoolean(env)") boolean cachedGuard) {
+            return false;
+        }
+
+        static boolean getBoolean(Env context) {
+            return context != null;
         }
     }