+++ truffle/src/com.oracle.truffle.dsl.processor/src/com/oracle/truffle/dsl/processor/parser/NodeParser.java
@@ -176,12 +176,8 @@ public final class NodeParser extends AbstractParser<NodeData> {
     }
 
     public static List<TypeMirror> getCachedAnnotations() {
-        ProcessorContext localContext = ProcessorContext.getInstance();
-        TypeMirror cacheAnnotation = localContext.getTypes().Cached;
-        TypeMirror cachedLibraryAnnotation = localContext.getTypes().CachedLibrary;
-        TypeMirror cachedContextAnnotation = localContext.getTypes().CachedContext;
-        TypeMirror cachedLanguageAnnotation = localContext.getTypes().CachedLanguage;
-        return Arrays.asList(cacheAnnotation, cachedLibraryAnnotation, cachedContextAnnotation, cachedLanguageAnnotation);
+        TruffleTypes types = ProcessorContext.getInstance().getTypes();
+        return Arrays.asList(types.Cached, types.CachedLibrary, types.CachedContext, types.CachedLanguage, types.Bind);
     }
 
     public static NodeParser createExportParser(TypeMirror exportLibraryType, TypeElement exportDeclarationType, boolean substituteThisToParent) {
@@ -709,15 +705,24 @@ public final class NodeParser extends AbstractParser<NodeData> {
 
             List<SpecializationData> failedSpecializations = null;
             if (specialization.isReachesFallback() && !specialization.getCaches().isEmpty() && !specialization.getGuards().isEmpty()) {
-                boolean guardBoundByCache = false;
-                for (GuardExpression guard : specialization.getGuards()) {
-                    if (specialization.isGuardBoundWithCache(guard)) {
-                        guardBoundByCache = true;
+                boolean failed = false;
+                if (specialization.getMaximumNumberOfInstances() > 1) {
+                    for (GuardExpression guard : specialization.getGuards()) {
+                        if (specialization.isGuardBoundWithCache(guard)) {
+                            failed = true;
+                            break;
+                        }
+                    }
+                }
+
+                for (CacheExpression cache : specialization.getCaches()) {
+                    if (cache.isGuardForNull()) {
+                        failed = true;
                         break;
                     }
                 }
 
-                if (guardBoundByCache && specialization.getMaximumNumberOfInstances() > 1) {
+                if (failed) {
                     if (failedSpecializations == null) {
                         failedSpecializations = new ArrayList<>();
                     }
@@ -727,16 +732,15 @@ public final class NodeParser extends AbstractParser<NodeData> {
 
             if (failedSpecializations != null) {
                 List<String> specializationIds = failedSpecializations.stream().map((e) -> e.getId()).collect(Collectors.toList());
-
                 fallback.addError(
                                 "Some guards for the following specializations could not be negated for the @%s specialization: %s. " +
-                                                "Guards cannot be negated for the @%s when they bind @%s parameters and the specialization may consist of multiple instances. " +
+                                                "Guards cannot be negated for the @%s when they bind @%s parameters and the specialization may consist of multiple instances or if any of the @%s parameters is configured as weak. " +
                                                 "To fix this limit the number of instances to '1' or " +
                                                 "introduce a more generic specialization declared between this specialization and the fallback. " +
                                                 "Alternatively the use of @%s can be avoided by declaring a @%s with manually specified negated guards.",
-                                types.Fallback.asElement().getSimpleName().toString(), specializationIds, types.Fallback.asElement().getSimpleName().toString(),
-                                types.Cached.asElement().getSimpleName().toString(), types.Fallback.asElement().getSimpleName().toString(),
-                                types.Specialization.asElement().getSimpleName().toString());
+                                ElementUtils.getSimpleName(types.Fallback), specializationIds,
+                                ElementUtils.getSimpleName(types.Fallback), ElementUtils.getSimpleName(types.Cached), ElementUtils.getSimpleName(types.Cached),
+                                ElementUtils.getSimpleName(types.Fallback), ElementUtils.getSimpleName(types.Specialization));
             }
 
         }
@@ -1960,7 +1964,9 @@ public final class NodeParser extends AbstractParser<NodeData> {
     private SpecializationData initializeCaches(SpecializationData specialization, DSLExpressionResolver resolver) {
         List<CacheExpression> caches = new ArrayList<>();
         List<CacheExpression> cachedLibraries = new ArrayList<>();
-        parameters: for (Parameter parameter : specialization.getParameters()) {
+
+        Parameter[] parameters = specialization.getParameters().toArray(new Parameter[0]);
+        parameters: for (Parameter parameter : parameters) {
             if (!parameter.getSpecification().isCached()) {
                 continue;
             }
@@ -1993,7 +1999,48 @@ public final class NodeParser extends AbstractParser<NodeData> {
             caches.add(cache);
 
             if (cache.isCached()) {
-                parseCached(cache, specialization, resolver, parameter);
+                boolean weakReference = getAnnotationValue(Boolean.class, foundCached, "weak");
+                if (weakReference) {
+                    if (ElementUtils.isPrimitive(cache.getParameter().getType())) {
+                        cache.addError("Cached parameters with primitive types cannot be weak. Set weak to false to resolve this.");
+                    }
+
+                    parseCached(cache, specialization, resolver, parameter);
+                    if (cache.hasErrors()) {
+                        continue;
+                    }
+
+                    DSLExpression sourceExpression = cache.getDefaultExpression();
+
+                    String weakName = "weak" + ElementUtils.firstLetterUpperCase(parameter.getLocalName()) + "Gen_";
+                    TypeMirror weakType = new CodeTypeMirror.DeclaredCodeTypeMirror(context.getTypeElement(types.TruffleWeakReference), Arrays.asList(cache.getParameter().getType()));
+                    CodeVariableElement weakVariable = new CodeVariableElement(weakType, weakName);
+                    Parameter weakParameter = new Parameter(parameter, weakVariable);
+
+                    DSLExpression newWeakReference = new DSLExpression.Call(null, "new", Arrays.asList(sourceExpression));
+                    newWeakReference.setResolvedTargetType(weakType);
+                    resolveCachedExpression(resolver, cache, weakType, newWeakReference, null);
+
+                    CacheExpression weakCache = new CacheExpression(weakParameter, foundCached);
+                    weakCache.setDefaultExpression(newWeakReference);
+                    weakCache.setUncachedExpression(newWeakReference);
+                    weakCache.setWeakReference(true);
+
+                    caches.add(0, weakCache);
+
+                    DSLExpressionResolver weakResolver = resolver.copy(Arrays.asList());
+                    weakResolver.addVariable(weakName, weakVariable);
+                    specialization.addParameter(specialization.getParameters().size(), weakParameter);
+
+                    DSLExpression parsedDefaultExpression = parseCachedExpression(weakResolver, cache, parameter.getType(), weakName + ".get()");
+                    cache.setDefaultExpression(parsedDefaultExpression);
+                    cache.setUncachedExpression(sourceExpression);
+                    cache.setAlwaysInitialized(true);
+                    cache.setGuardForNull(true);
+                } else {
+                    parseCached(cache, specialization, resolver, parameter);
+                }
+
             } else if (cache.isCachedLibrary()) {
                 AnnotationMirror cachedLibrary = cache.getMessageAnnotation();
                 String expression = getCachedLibraryExpressions(cachedLibrary);
@@ -2129,6 +2176,14 @@ public final class NodeParser extends AbstractParser<NodeData> {
                 cache.setLanguageType(languageType);
                 cache.setDefaultExpression(resolveCachedExpression(resolver, cache, null, accessReference, null));
                 cache.setUncachedExpression(resolveCachedExpression(resolver, cache, null, accessReference, null));
+                cache.setAlwaysInitialized(true);
+            } else if (cache.isBind()) {
+                AnnotationMirror dynamic = cache.getMessageAnnotation();
+                String expression = ElementUtils.getAnnotationValue(String.class, dynamic, "value", false);
+
+                DSLExpression parsedExpression = parseCachedExpression(resolver, cache, parameter.getType(), expression);
+                cache.setDefaultExpression(parsedExpression);
+                cache.setUncachedExpression(parsedExpression);
                 cache.setAlwaysInitialized(true);
             }
         }
@@ -2374,6 +2429,7 @@ public final class NodeParser extends AbstractParser<NodeData> {
         }
 
         List<String> expressionParameters = getAnnotationValueList(String.class, cachedAnnotation, "parameters");
+
         String initializer = getAnnotationValue(String.class, cachedAnnotation, "value");
         String uncached = getAnnotationValue(String.class, cachedAnnotation, "uncached");
 
@@ -2441,9 +2497,7 @@ public final class NodeParser extends AbstractParser<NodeData> {
                         cache.getMessages().clear();
                     }
                 }
-
             }
-
         }
 
         if (requireUncached && cache.getUncachedExpression() == null && cache.getDefaultExpression() != null) {
@@ -2451,8 +2505,6 @@ public final class NodeParser extends AbstractParser<NodeData> {
                 cache.setUncachedExpression(cache.getDefaultExpression());
             }
         }
-
-        return;
     }
 
     private DSLExpression resolveCachedExpression(DSLExpressionResolver resolver, CacheExpression cache, TypeMirror targetType, DSLExpression expression, String originalString) {
@@ -2492,24 +2544,62 @@ public final class NodeParser extends AbstractParser<NodeData> {
     }
 
     private void initializeGuards(SpecializationData specialization, DSLExpressionResolver resolver) {
-        final TypeMirror booleanType = context.getType(boolean.class);
         List<String> guardDefinitions = getAnnotationValueList(String.class, specialization.getMarkerAnnotation(), "guards");
-        for (String guard : guardDefinitions) {
-            GuardExpression guardExpression;
-            DSLExpression expression = null;
-            try {
-                expression = DSLExpression.parse(guard);
-                expression.accept(resolver);
-                guardExpression = new GuardExpression(specialization, expression);
-                if (!typeEquals(expression.getResolvedType(), booleanType)) {
-                    guardExpression.addError("Incompatible return type %s. Guards must return %s.", getSimpleName(expression.getResolvedType()), getSimpleName(booleanType));
+
+        Set<CacheExpression> handledCaches = new HashSet<>();
+        List<GuardExpression> guards = new ArrayList<>();
+        for (String guardExpression : guardDefinitions) {
+            GuardExpression guard = parseGuard(resolver, specialization, guardExpression);
+
+            if (guard.getExpression() != null) {
+                Set<CacheExpression> caches = specialization.getBoundCaches(guard.getExpression(), false);
+                for (CacheExpression cache : caches) {
+                    if (handledCaches.contains(cache)) {
+                        continue;
+                    }
+                    if (cache.isGuardForNull()) {
+                        guards.add(createWeakReferenceGuard(resolver, specialization, cache));
+                    }
                 }
-            } catch (InvalidExpressionException e) {
-                guardExpression = new GuardExpression(specialization, null);
-                guardExpression.addError("Error parsing expression '%s': %s", guard, e.getMessage());
+                handledCaches.addAll(caches);
             }
-            specialization.getGuards().add(guardExpression);
+
+            guards.add(guard);
         }
+        for (CacheExpression cache : specialization.getCaches()) {
+            if (cache.isGuardForNull()) {
+                if (handledCaches.contains(cache)) {
+                    continue;
+                }
+                guards.add(createWeakReferenceGuard(resolver, specialization, cache));
+            }
+        }
+
+        specialization.getGuards().addAll(guards);
+    }
+
+    private GuardExpression createWeakReferenceGuard(DSLExpressionResolver resolver, SpecializationData specialization, CacheExpression cache) {
+        GuardExpression guard = parseGuard(resolver, specialization, cache.getParameter().getLocalName() + " != null");
+        guard.setWeakReferenceGuard(true);
+        return guard;
+    }
+
+    private GuardExpression parseGuard(DSLExpressionResolver resolver, SpecializationData specialization, String guard) {
+        final TypeMirror booleanType = context.getType(boolean.class);
+        GuardExpression guardExpression;
+        DSLExpression expression;
+        try {
+            expression = DSLExpression.parse(guard);
+            expression.accept(resolver);
+            guardExpression = new GuardExpression(specialization, expression);
+            if (!typeEquals(expression.getResolvedType(), booleanType)) {
+                guardExpression.addError("Incompatible return type %s. Guards must return %s.", getSimpleName(expression.getResolvedType()), getSimpleName(booleanType));
+            }
+        } catch (InvalidExpressionException e) {
+            guardExpression = new GuardExpression(specialization, null);
+            guardExpression.addError("Error parsing expression '%s': %s", guard, e.getMessage());
+        }
+        return guardExpression;
     }
 
     private void initializeGeneric(final NodeData node) {